Let claude bring the iOS app to run again in sim

+++ code/osx/Blinkensim/Stereoscope/Engine/Renderer/Core/GraphicsDevice/Texture.mm
@@ -23,7 +23,7 @@ subject to the following restrictions:
 #include "Resource.h"
 #include "Macros.h"
 
-typedef unsigned long U32;
+typedef unsigned int U32;
 typedef unsigned char U8;
 
 //DEFINE_HEAP(CTexture, "Textures");
@@ -1322,10 +1322,10 @@ const int mod[8][4]={{2, 8,-2,-8},
 
 
 // lsb: hgfedcba ponmlkji msb: hgfedcba ponmlkji due to endianness
-unsigned long modifyPixel(int red, int green, int blue, int x, int y, unsigned long modBlock, int modTable)
+unsigned int modifyPixel(int red, int green, int blue, int x, int y, unsigned int modBlock, int modTable)
 {
 	int index = x*4+y, pixelMod;
-	unsigned long mostSig = modBlock<<1;
+	unsigned int mostSig = modBlock<<1;
 	if (index<8)	//hgfedcba
 		pixelMod = mod[modTable][((modBlock>>(index+24))&0x1)+((mostSig>>(index+8))&0x2)];
 	else	// ponmlkj
@@ -1350,8 +1350,8 @@ unsigned long modifyPixel(int red, int green, int blue, int x, int y, unsigned l
 //					
 int ETCTextureDecompress(const void * const pSrcData, const int &x, const int &y, const void *pDestData,const int &nMode)
 {
-	unsigned long blockTop, blockBot, *input = (unsigned long*)pSrcData, *output;
-	//unsigned long alphaTop, alphaBot, alphaValue[8];
+	unsigned int blockTop, blockBot, *input = (unsigned int*)pSrcData, *output;
+	//unsigned int alphaTop, alphaBot, alphaValue[8];
 	unsigned char red1, green1, blue1, red2, green2, blue2;
 	bool bFlip, bDiff;
 	int modtable1,modtable2;
@@ -1363,7 +1363,7 @@ int ETCTextureDecompress(const void * const pSrcData, const int &x, const int &y
 				blockTop = *(input++);
 				blockBot = *(input++);
 
-			output = (unsigned long*)pDestData + i*x +m;
+			output = (unsigned int*)pDestData + i*x +m;
 				// check flipbit
 			bFlip = (blockTop & ETC_FLIP) != 0;
 			bDiff = (blockTop & ETC_DIFF) != 0;
